[arc] - simplified RefreshInputs algorithm

+++ cesium/delete_test.go
@@ -533,13 +533,13 @@ var _ = Describe("Delete", func() {
 						Expect(frame.Count()).To(Equal(2))
 						Expect(frame.SeriesAt(0).TimeRange.End).To(Equal(12 * telem.SecondTS))
 
-						series0Data := telem.UnmarshalSlice[int](frame.SeriesAt(0).Data, telem.Int64T)
+						series0Data := telem.UnmarshalSlice[int64](frame.SeriesAt(0).Data, telem.Int64T)
 						Expect(series0Data).To(ContainElement(0))
 						Expect(series0Data).To(ContainElement(1))
 						Expect(series0Data).ToNot(ContainElement(2))
 
 						Expect(frame.SeriesAt(1).TimeRange.Start).To(Equal(17 * telem.SecondTS))
-						series1Data := telem.UnmarshalSlice[int](frame.SeriesAt(1).Data, telem.Int64T)
+						series1Data := telem.UnmarshalSlice[int64](frame.SeriesAt(1).Data, telem.Int64T)
 
 						Expect(series1Data).ToNot(ContainElement(6))
 						Expect(series1Data).To(ContainElement(7))
